Rename classes and move member function definitions.
