+++ alvr_server/alvr_server.vcxproj
@@ -109,6 +109,8 @@
     <ClCompile Include="ControlSocket.cpp" />
     <ClCompile Include="alvr_server.cpp" />
     <ClCompile Include="DeviceQuery.cpp" />
+    <ClCompile Include="OpenVRController.cpp" />
+    <ClCompile Include="OpenVRFakeTrackingReference.cpp" />
     <ClCompile Include="FrameEncoder.cpp" />
     <ClCompile Include="FrameRender.cpp" />
     <ClCompile Include="FreePIE.cpp" />
@@ -125,9 +127,10 @@
     <ClCompile Include="openvr-utils\threadtools.cpp" />
     <ClCompile Include="OpenVRDirectModeComponent.cpp" />
     <ClCompile Include="OpenVRDisplayComponent.cpp" />
-    <ClCompile Include="OpenVRServerDriver.cpp" />
+    <ClCompile Include="OpenVRHmd.cpp" />
     <ClCompile Include="PacketBuffer.cpp" />
     <ClCompile Include="Poller.cpp" />
+    <ClCompile Include="RecenterManager.cpp" />
     <ClCompile Include="Settings.cpp" />
     <ClCompile Include="ThrottlingBuffer.cpp" />
     <ClCompile Include="UdpSocket.cpp" />
@@ -200,17 +203,17 @@
     <ClInclude Include="openvr-utils\threadtools.h" />
     <ClInclude Include="OpenVRDirectModeComponent.h" />
     <ClInclude Include="OpenVRDisplayComponent.h" />
-    <ClInclude Include="OpenVRServerDriver.h" />
+    <ClInclude Include="OpenVRHmd.h" />
     <ClInclude Include="PacketBuffer.h" />
     <ClInclude Include="Poller.h" />
     <ClInclude Include="RecenterManager.h" />
-    <ClInclude Include="RemoteController.h" />
+    <ClInclude Include="OpenVRController.h" />
     <ClInclude Include="ResampleUtils.h" />
     <ClInclude Include="resource.h" />
     <ClInclude Include="Settings.h" />
     <ClInclude Include="Statistics.h" />
     <ClInclude Include="ThrottlingBuffer.h" />
-    <ClInclude Include="Tracking.h" />
+    <ClInclude Include="OpenVRFakeTrackingReference.h" />
     <ClInclude Include="UdpSocket.h" />
     <ClInclude Include="Utils.h" />
     <ClInclude Include="VideoEncoder.h" />
